lint

+++ src/crab/ebpf_domain.cpp
@@ -392,8 +392,8 @@ EbpfDomain EbpfDomain::from_constraints(const std::set<std::string>& constraints
     }
     for (const Interval& range : numeric_ranges) {
         const auto [start, ub] = range.pair<int64_t>();
-        const int width = 1 + (ub - start);
-        inv.stack.initialize_numbers(start, width);
+        const int width = gsl::narrow<int>(1 + (ub - start));
+        inv.stack.initialize_numbers(gsl::narrow<int>(start), width);
     }
     // TODO: handle other stack type constraints
     return inv;
